+++ lib/dot_grid/pattern/checkerboard.rb
@@ -1,27 +1,6 @@
 module DotGrid
   module Pattern
-    class Checkerboard
-      attr_accessor(
-        :pdf,
-        :bounds,
-        :grid_color,
-        :spacing,
-      )
-
-      def initialize(params = {})
-        @pdf = params[:pdf]
-        @bounds = params[:bounds]
-        @grid_color = params[:grid_color] || "B3B3B3"
-        @spacing = params[:spacing] ? params[:spacing].mm : 5.mm
-      end
-
-      def rows
-        (bounds.height / spacing).floor
-      end
-
-      def columns
-        (bounds.width / spacing).floor
-      end
+    class Checkerboard < Pattern
 
       def draw
         pdf.fill_color grid_color
