+++ lib/dot_grid/pattern/horizontal_rule.rb
@@ -1,23 +1,6 @@
 module DotGrid
   module Pattern
-    class HorizontalRule
-      attr_accessor(
-        :pdf,
-        :bounds,
-        :grid_color,
-        :spacing,
-      )
-
-      def initialize(params = {})
-        @pdf = params[:pdf]
-        @bounds = params[:bounds]
-        @grid_color = params[:grid_color] || "B3B3B3"
-        @spacing = params[:spacing] ? params[:spacing].mm : 5.mm
-      end
-
-      def rows
-        (bounds.height / spacing).floor
-      end
+    class HorizontalRule < Pattern
 
       def draw
         pdf.stroke_color grid_color
