+++ spec/lib/dot_grid/pattern/horizontal_rule_spec.rb
@@ -1,17 +1,12 @@
 require 'spec_helper'
 
 describe "DotGrid::Pattern::HorizontalRule" do
-  let(:pdf) { double('pdf') }
-  let(:params) { { pdf: pdf, bounds: double('bounds', { height: 20.mm, width: 10.mm } ), spacing: 5 } }
-  let(:subject) { DotGrid::Pattern::HorizontalRule.new(params) }
-
-  describe "#rows" do
-    it "calculates the rows based on the bounds" do
-      expect(subject.rows).to eq(4)
-    end
-  end
 
   describe "#draw" do
+    let(:pdf) { double('pdf') }
+    let(:params) { { pdf: pdf, bounds: double('bounds', { height: 20.mm, width: 10.mm } ), spacing: 5 } }
+    let(:subject) { DotGrid::Pattern::HorizontalRule.new(params) }
+
     before do
       allow(pdf).to receive(:stroke_color)
       allow(pdf).to receive(:stroke_horizontal_line)
@@ -29,5 +24,3 @@ describe "DotGrid::Pattern::HorizontalRule" do
     end
   end
 end
-
-
